renamed file to concurrent, fixed test
