+++ tests/test_snql_api.py
@@ -78,10 +78,6 @@ class TestSnQLApi(BaseApiTest):
         )
 
     def test_simple_query(self) -> None:
-        from snuba import state
-
-        state.set_config("project_referrer_concurrent_limit_myreferrer", 1)
-        state.set_config("project_concurrent_limit", 0)
         response = self.post(
             "/discover/snql",
             data=json.dumps(
